requirements.txt
langchain>=0.2.0
langchain-huggingface>=0.0.5
huggingface_hub>=0.21.4
pydantic>=2.0
tqdm>=4.66.4
python-dotenv>=1.0.1
openai>=1.30.1  # Optional, ignore if you're not using OpenAI
